% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCalcLengthOfGaps.R
\name{fCalcLengthOfGaps}
\alias{fCalcLengthOfGaps}
\title{Calculate length of gaps (with flag NA) in specified vector}
\usage{
fCalcLengthOfGaps(Data.V.n)
}
\arguments{
\item{Data.V.n}{Numeric vector with gaps (missing values, NAs)}
}
\value{
VA numeric value - number of missing values
}
\description{
Calculate length of gaps (with flag NA) in specified vector
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
