% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCheckColNames.R
\name{fCheckColNames}
\alias{fCheckColNames}
\title{Check if specified columns exists in data frame}
\usage{
fCheckColNames(Data.F, ColNames.V.s)
}
\arguments{
\item{Data.F}{Data frame}

\item{ColNames.V.s}{Character Vector of variables to be checked}
}
\value{
Nothing. Function stops on error
}
\description{
Check and set quality flag
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
