% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCheckValNum.R
\name{fCheckValNum}
\alias{fCheckValNum}
\title{Check if variable is a numeric}
\usage{
fCheckValNum(Value.n)
}
\arguments{
\item{Value.n}{Value to be checked if numeric (but can also be NA of any type)}
}
\value{
Logical value indicating variable is numeric (TRUE) or not-numeric (FALSE)
}
\description{
Check if variable is a numeric
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
