% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCheckValString.R
\name{fCheckValString}
\alias{fCheckValString}
\title{Check if variable is a non-empty character string}
\usage{
fCheckValString(Value.s)
}
\arguments{
\item{Value.s}{Value to be checked if string}
}
\value{
Logical value indicating variable is string (TRUE) or not-string (FALSE)
}
\description{
Check if variable is a non-empty character string
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
