% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fConvertGapsToNA.R
\name{fConvertGapsToNA}
\alias{fConvertGapsToNA}
\title{Converts all gap flags to NA}
\usage{
fConvertGapsToNA(Data.F, GapFlag.n = -9999)
}
\arguments{
\item{Data.F}{Data frame}

\item{GapFlag.n}{Flag value used for gaps, defaults to -9999.0}
}
\value{
Data.F with gap flags turned to NA
}
\description{
Converts all gap flags to NA
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
