% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fSetQF.R
\name{fSetQF}
\alias{fSetQF}
\title{Check and set quality flag}
\usage{
fSetQF(Data.F, Var.s, QFVar.s, QFValue.n)
}
\arguments{
\item{Data.F}{Data frame}

\item{Var.s}{Variable to be filled}

\item{QFVar.s}{Quality flag of variable to be filled}

\item{QFValue.n}{Value of quality flag for _good_ (original) data, other data is set to missing}
}
\value{
Var.s with quality flags applied
}
\description{
Check and set quality flag
}
\examples{
Currently None
}
\references{
None
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
