% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFillInit.R
\name{sFillInit}
\alias{sFillInit}
\title{Initialize gap-filling output}
\usage{
sFillInit(sDATA, Var.s, QFVar.s = "none", QFValue.n = NA_real_,
  FillAll.b = TRUE)
}
\arguments{
\item{sDATA}{Data frame}

\item{Var.s}{Variable to be filled}

\item{QFVar.s}{Quality flag of variable to be filled}

\item{QFValue.n}{Value of quality flag for _good_ (original) data, other data is set to missing}

\item{FillAll.b}{Fill all values to estimate uncertainties?}
}
\value{
A list of:\cr
Data frame sTEMP with:
VAR\emph{_orig} - Original values used for gap filling \cr
VAR\emph{_f   } - Original values and gaps filled with mean of selected datapoints (condition depending on gap filling method) \cr
VAR\emph{_fqc} - Quality flag assigned depending on gap filling method and window length (0 = original data, 1 = most reliable, 2 = medium, 3 = least reliable) \cr
VAR\emph{_fall} - All values considered as gaps (for uncertainty estimates) \cr
VAR\emph{_fall_qc} - Quality flag assigned depending on gap filling method and window length (1 = most reliable, 2 = medium, 3 = least reliable) \cr
VAR\emph{_fnum} - Number of datapoints used for gap-filling \cr
VAR\emph{_fsd} - Standard deviation of datapoints used for gap filling (uncertainty) \cr
VAR\emph{_fmeth} - Method used for gap filling (1 = similar meteo condition (sFillLUT with Rg, VPD, Tair), 2 = similar meteo (sFillLUT with Rg only), 3 = mean diurnal course (sFillMDC)) \cr
VAR\emph{_fwin} - Full window length used for gap filling \cr
\cr
List sINFO
}
\description{
initialize Gap-filling output for gap-filling method after Reichstein et al. (2005)
}
\examples{
Currently None
}
\references{
M. Reichstein, E. Falge, D. Baldocchi, D. Papale, M. Aubinet, P. Berbigier, C. Bernhofer, N. Buchmann, T. Gilmanov, A. Granier, T. Grunwald, K. Havrankova, H. Ilvesniemi, D. Janous, A. Knohl, T. Laurila, A. Lohila, D. Loustau, G. Matteucci, T. Meyers, F. Miglietta, J.M. Ourcival, J. Pumpanen, S. Rambal, E. Rotenberg, M. Sanz, J. Tenhunen, G. Seufert, F. Vaccari, T. Vesala, D. Yakir, R. Valentini: On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm, Global Change Biology 11(9) 1424–1439 (2005)
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
