% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFillLUT.R
\name{sFillLUT}
\alias{sFillLUT}
\title{Gap-fill with Look up Table}
\usage{
sFillLUT(sTEMP, sINFO, WinDays.i, V1.s = "none", T1.n = NA_real_,
  V2.s = "none", T2.n = NA_real_, V3.s = "none", T3.n = NA_real_,
  V4.s = "none", T4.n = NA_real_, V5.s = "none", T5.n = NA_real_,
  Verbose.b = TRUE)
}
\arguments{
\item{sTEMP}{# Initialized data frame sTEMP for newly generated gap filled data and qualifiers, as produced by sFillInit}

\item{sINFO}{# Information on gap-filling dataset, as produced by sFillInit}

\item{WinDays.i}{Numeric. Window size for filling [days]}

\item{V1.s}{Condition variable 1}

\item{T1.n}{Tolerance interval 1}

\item{V2.s}{}

\item{T2.n}{}

\item{V3.s}{}

\item{T3.n}{}

\item{V4.s}{}

\item{T4.n}{}

\item{V5.s}{}

\item{T5.n}{}
}
\value{
A data frame of:\cr
\code{'VAR_orig','VAR_f','VAR_fall','VAR_fnum','VAR_fsd','VAR_fwin'}
}
\description{
Gap-fill using Look up Table, after Reichstein et al. (2005)
}
\examples{
Currently None
}
\references{
M. Reichstein, E. Falge, D. Baldocchi, D. Papale, M. Aubinet, P. Berbigier, C. Bernhofer, N. Buchmann, T. Gilmanov, A. Granier, T. Grunwald, K. Havrankova, H. Ilvesniemi, D. Janous, A. Knohl, T. Laurila, A. Lohila, D. Loustau, G. Matteucci, T. Meyers, F. Miglietta, J.M. Ourcival, J. Pumpanen, S. Rambal, E. Rotenberg, M. Sanz, J. Tenhunen, G. Seufert, F. Vaccari, T. Vesala, D. Yakir, R. Valentini: On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm, Global Change Biology 11(9) 1424–1439 (2005)
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
