% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFillMDC.R
\name{sFillMDC}
\alias{sFillMDC}
\title{Gap-fill with Mean Diurnal Course method}
\usage{
sFillMDC(sTEMP, sINFO, WinDays.i, Verbose.b = TRUE)
}
\arguments{
\item{sTEMP}{# Initialized data frame sTEMP for newly generated gap filled data and qualifiers, as produced by sFillInit}

\item{sINFO}{# Information on gap-filling dataset, as produced by sFillInit}

\item{WinDays.i}{Window size for filling in days}

\item{Verbose.b}{Logical. Print status information to screen?}
}
\value{
A data frame of:\cr
}
\description{
Gap-fill with Mean Diurnal Course (MDC) algorithm based on average values within +/- one hour of adjacent days, after 
procedure in Reichstein et al. (2005)
}
\examples{
Currently None
}
\references{
M. Reichstein, E. Falge, D. Baldocchi, D. Papale, M. Aubinet, P. Berbigier, C. Bernhofer, N. Buchmann, T. Gilmanov, A. Granier, T. Grunwald, K. Havrankova, H. Ilvesniemi, D. Janous, A. Knohl, T. Laurila, A. Lohila, D. Loustau, G. Matteucci, T. Meyers, F. Miglietta, J.M. Ourcival, J. Pumpanen, S. Rambal, E. Rotenberg, M. Sanz, J. Tenhunen, G. Seufert, F. Vaccari, T. Vesala, D. Yakir, R. Valentini: On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm, Global Change Biology 11(9) 1424–1439 (2005)
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
