% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMDSGapFill.R
\name{sMDSGapFill}
\alias{sMDSGapFill}
\title{MDS gap filling algorithm adapted after the PV-Wave code and paper by Markus Reichstein.}
\usage{
sMDSGapFill(sDATA, Var.s, QFVar.s = "none", QFValue.n = NA_real_,
  V1.s = "Rg", T1.n = 50, V2.s = "VPD", T2.n = 5, V3.s = "Tair",
  T3.n = 2.5, FillAll.b = TRUE, Verbose.b = TRUE, Suffix.s = "")
}
\arguments{
\item{sDATA}{Named data frame with all variables indicated below as well as variable 'sDateTime' in POSIXct format}

\item{Var.s}{Character. Name of Variable to be filled}

\item{QFVar.s}{Character. Name of Quality flag of variable to be filled}

\item{QFValue.n}{Value of quality flag for _good_ (original) data, other data is set to missing}

\item{V1.s}{Character. Name of condition variable 1}

\item{T1.n}{Tolerance interval 1}

\item{V2.s}{Character. Name of condition variable 2}

\item{T2.n}{Tolerance interval 2}

\item{V3.s}{Character. Name of condition variable 3}

\item{T3.n}{Tolerance interval 3}

\item{FillAll.b}{Logical. Fill all values to estimate uncertainties?}

\item{Verbose.b}{Logical. Print status information to screen}

\item{Suffix.s}{String suffix needed for different processing setups on the same dataset (for explanations see below)}
}
\value{
A data frame of:\cr
VAR\emph{_orig} - Original values used for gap filling \cr
VAR\emph{_f   } - Original values and gaps filled with mean of selected datapoints (condition depending on gap filling method) \cr
VAR\emph{_fqc} - Quality flag assigned depending on gap filling method and window length (0 = original data, 1 = most reliable, 2 = medium, 3 = least reliable) \cr
VAR\emph{_fall} - All values considered as gaps (for uncertainty estimates) \cr
VAR\emph{_fall_qc} - Quality flag assigned depending on gap filling method and window length (1 = most reliable, 2 = medium, 3 = least reliable) \cr
VAR\emph{_fnum} - Number of datapoints used for gap-filling \cr
VAR\emph{_fsd} - Standard deviation of datapoints used for gap filling (uncertainty) \cr
VAR\emph{_fmeth} - Method used for gap filling (1 = similar meteo condition (sFillLUT with Rg, VPD, Tair), 2 = similar meteo (sFillLUT with Rg only), 3 = mean diurnal course (sFillMDC)) \cr
VAR\emph{_fwin} - Full window length used for gap filling \cr
}
\description{
MDS gap filling algorithm adapted after the PV-Wave code and paper by Markus Reichstein (2005).
}
\examples{
Currently None
}
\references{
M. Reichstein, E. Falge, D. Baldocchi, D. Papale, M. Aubinet, P. Berbigier, C. Bernhofer, N. Buchmann, T. Gilmanov, A. Granier, T. Grunwald, K. Havrankova, H. Ilvesniemi, D. Janous, A. Knohl, T. Laurila, A. Lohila, D. Loustau, G. Matteucci, T. Meyers, F. Miglietta, J.M. Ourcival, J. Pumpanen, S. Rambal, E. Rotenberg, M. Sanz, J. Tenhunen, G. Seufert, F. Vaccari, T. Vesala, D. Yakir, R. Valentini: On the separation of net ecosystem exchange into assimilation and ecosystem respiration: review and improved algorithm, Global Change Biology 11(9) 1424–1439 (2005)
}
\author{
AMM
}
\keyword{Currently}
\keyword{None.}
